[refactor] PenDetail 相關元件位置整理 [feat] PenDetailContent 新增作品描述與標籤區塊
