--- src/components/PenDetailContent.vue
@@ -1,15 +1,0 @@
-<script setup>
-import { defineProps } from "vue";
-import CommentList from "@/components/Comments/CommentList.vue";
-const props = defineProps({
-  pen: {
-    type: Object,
-    required: true,
-  },
-});
-</script>
-<template>
-  <div>
-    <CommentList :penId="pen.id" @close="$emit('close')"/>
-  </div>
-</template>
