+++ src/components/PenDetails/PenDetailModalHeader.vue
