+++ src/components/PenDetails/PenDetailPageHeader.vue
